Stroke out functional approach for CachedContent CURD ops
